refactor: separate time in a file
